tests/wpt/web-platform-tests/css/css-sizing/table-percentage-min-width-below-float.html

<!DOCTYPE html>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-sizing/#available">
<link rel="help" href="https://drafts.csswg.org/css-sizing/#percentage-sizing">
<p>There should be a blue square below.</p>
<div id="container" style="width:100px;">
  <div style="float:right; width:50px; height:10px;"></div>
  <div style="display:table; width:10px; min-width:100%; height:100px; background:blue;" data-expected-width="100"></div>
</div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script>
  checkLayout("#container");
</script>
